+++ CHANGELOG.md
@@ -32,6 +32,7 @@
 - Dev: Don't use `stampede` for link resolver links. (#394)
 - Dev: Update to Twitter's v2 API. (#414)
 - Dev: Add HTTP Caching headers. (#417)
+- Dev: Add custom middleware to strip trailing slashes. (#422)
 - Dev: Make cache timeout durations configurable. (#419)
 - Dev: Add prometheus middleware for chi. (#420)